requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
transformers>=4.35.0
timm>=0.9.0
albumentations>=1.3.1
opencv-python-headless>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tensorboard>=2.14.0
PyYAML>=6.0
tqdm>=4.66.0
pandas>=2.0.0
ttach>=0.0.3
